+++ src/main/java/org/minnnisu/togetherdelivery/controller/ChatMessageController.java
@@ -3,9 +3,9 @@ package org.minnnisu.togetherdelivery.controller;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.minnnisu.togetherdelivery.domain.User;
-import org.minnnisu.togetherdelivery.dto.chat.chatMessage.chatMessageDelete.ChatMessageDeleteDto;
-import org.minnnisu.togetherdelivery.dto.chat.chatMessage.chatMessageDelete.ChatMessageDeleteResponseDto;
-import org.minnnisu.togetherdelivery.dto.chat.stompChatMessage.StompChatMessageDeleteResponseDto;
+import org.minnnisu.togetherdelivery.dto.stomp.chatMessage.chatMessageDelete.ChatMessageDeleteDto;
+import org.minnnisu.togetherdelivery.dto.stomp.chatMessage.chatMessageDelete.ChatMessageDeleteResponseDto;
+import org.minnnisu.togetherdelivery.dto.stomp.stompChatMessage.StompChatMessageDeleteResponseDto;
 import org.minnnisu.togetherdelivery.service.ChatMessageService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
