+++ src/main/java/org/minnnisu/togetherdelivery/dto/stomp/stompChatMessage/StompChatMessageLeaveResponseDto.java
@@ -1,4 +1,4 @@
-package org.minnnisu.togetherdelivery.dto.chat.stompChatMessage;
+package org.minnnisu.togetherdelivery.dto.stomp.stompChatMessage;
 
 import lombok.Getter;
 import lombok.Setter;
