+++ src/main/java/org/minnnisu/togetherdelivery/service/StompChatService.java
@@ -8,9 +8,8 @@ import org.minnnisu.togetherdelivery.domain.ChatMessage;
 import org.minnnisu.togetherdelivery.domain.ChatRoom;
 import org.minnnisu.togetherdelivery.domain.ChatRoomMember;
 import org.minnnisu.togetherdelivery.domain.User;
-import org.minnnisu.togetherdelivery.dto.chat.chatMessage.chatMessageInvite.ChatMessageDto;
-import org.minnnisu.togetherdelivery.dto.chat.chatMessage.stompChatMessage.*;
-import org.minnnisu.togetherdelivery.dto.chat.stompChatMessage.*;
+import org.minnnisu.togetherdelivery.dto.stomp.chatMessage.ChatMessageDto;
+import org.minnnisu.togetherdelivery.dto.stomp.stompChatMessage.*;
 import org.minnnisu.togetherdelivery.exception.CustomErrorException;
 import org.minnnisu.togetherdelivery.handler.AssignPrincipalHandshakeHandler.StompPrincipal;
 import org.minnnisu.togetherdelivery.repository.ChatMessageRepository;
