+++ src/test/java/org/minnnisu/togetherdelivery/controller/ChatRoomControllerTest.java
@@ -15,8 +15,8 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.minnnisu.togetherdelivery.constant.ChatMessageType;
 import org.minnnisu.togetherdelivery.constant.ErrorCode;
-import org.minnnisu.togetherdelivery.dto.chat.chatMessage.chatMessageInvite.ChatMessageDto;
-import org.minnnisu.togetherdelivery.dto.chat.stompChatMessage.StompChatMessageEnterResponseDto;
+import org.minnnisu.togetherdelivery.dto.stomp.chatMessage.ChatMessageDto;
+import org.minnnisu.togetherdelivery.dto.stomp.stompChatMessage.StompChatMessageEnterResponseDto;
 import org.minnnisu.togetherdelivery.dto.chat.chatRoom.chatRoomExit.ChatRoomExitRequestDto;
 import org.minnnisu.togetherdelivery.dto.chat.chatRoom.chatRoomExit.ChatRoomExitResponseDto;
 import org.minnnisu.togetherdelivery.dto.chat.chatRoom.chatRoomInvite.ChatRoomInviteDto;
